actualización de codigo
actualización de la clase excepción, incidenciaTic, pom,  Dtolncidencia

+++ ReaktorIssuesServer/src/main/java/es/iesjandula/ReaktorIssuesServer/repository/IncidenciaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import es.iesjandula.ReaktorIssuesServer.models.IncidenciaTic;
 
 
-public interface IncidenciaRepository extends JpaRepository<IncidenciaTic, Long>{
+public interface IncidenciaRepository extends JpaRepository<IncidenciaTic, Integer>{
 
 }
